move foreground qt
